+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/CoreWeb/AppSettings/AppSettingsTemplate.cs
@@ -34,8 +34,21 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
         {
             this.Write(" \r\n");
             this.Write(" \r\n");
-            this.Write("{\r\n  \"Logging\": {\r\n    \"LogLevel\": {\r\n      \"Default\": \"Information\",\r\n      \"Mic" +
-                    "rosoft.AspNetCore\": \"Warning\"\r\n    }\r\n  },\r\n  \"AllowedHosts\": \"*\"\r\n}\r\n");
+            this.Write("{\r\n  \"Logging\": {\r\n    \"LogLevel\": {\r\n      \"Default\": \"Warning\"\r\n    }\r\n  }");
+            
+            #line 17 "C:\Dev\Intent.Modules.NET\Modules\Intent.Modules.VisualStudio.Projects\Templates\CoreWeb\AppSettings\AppSettingsTemplate.tt"
+if (IncludeAllowHosts()) {
+            
+            #line default
+            #line hidden
+            this.Write(", \r\n  \"AllowedHosts\": \"*\"\r\n");
+            
+            #line 19 "C:\Dev\Intent.Modules.NET\Modules\Intent.Modules.VisualStudio.Projects\Templates\CoreWeb\AppSettings\AppSettingsTemplate.tt"
+ }
+            
+            #line default
+            #line hidden
+            this.Write("}\r\n");
             return this.GenerationEnvironment.ToString();
         }
     }