+++ Modules/Intent.Modules.WindowsServiceHost/Templates/ProgramFile.cs
@@ -1,0 +1,219 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Security.Cryptography;
+using Intent.Modules.Common.CSharp.AppStartup;
+using Intent.Modules.Common.CSharp.Builder;
+using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.CSharp.VisualStudio;
+using Intent.Modules.VisualStudio.Projects.Api;
+
+namespace Intent.Modules.WindowsServiceHost.Templates
+{
+    using static IProgramFile;
+
+    internal class ProgramFile : IProgramFile
+    {
+        private readonly CSharpTemplateBase<object> _template;
+        private readonly CSharpFile _cSharpFile;
+        private readonly bool _usesTopLevelStatements;
+
+        public ProgramFile(CSharpTemplateBase<object> template)
+        {
+            if (template is not ICSharpFileBuilderTemplate fileBuilderTemplate)
+            {
+                throw new Exception($"template must implement {nameof(ICSharpFileBuilderTemplate)}");
+            }
+
+            _template = template;
+            _cSharpFile = fileBuilderTemplate.CSharpFile;
+            UsesMinimalHostingModel = template.OutputTarget.GetProject()?.InternalElement?.AsCSharpProjectNETModel()?.GetNETSettings()?.UseMinimalHostingModel() == true;
+            _usesTopLevelStatements = _template.OutputTarget.GetProject()?.InternalElement?.AsCSharpProjectNETModel()?.GetNETSettings()?.UseTopLevelStatements() == true;
+        }
+
+        public bool UsesMinimalHostingModel { get; }
+
+        //public IProgramFile AddHostBuilderLambda(string methodName, CSharpLambdaBlock lambda)
+        //{
+        //    if (!UsesMinimalHostingModel)
+        //    {
+        //        _cSharpFile.OnBuild(file =>
+        //        {
+        //            var hostBuilder =
+        //                (IHasCSharpStatements)file.TopLevelStatements?.FindMethod("CreateHostBuilder") ??
+        //                file.Classes.First().FindMethod("CreateHostBuilder");
+
+        //            var hostBuilderChain = (CSharpMethodChainStatement)hostBuilder.Statements.First();
+        //            hostBuilderChain.Statements.Last().InsertAbove(new CSharpInvocationStatement(methodName)
+        //                .WithoutSemicolon()
+        //                .AddArgument(lambda));
+        //        });
+        //    }
+        //    else
+        //    {
+        //        _cSharpFile.OnBuild(file =>
+        //        {
+        //            var targetBlock = _usesTopLevelStatements
+        //                ? (IHasCSharpStatements)file.TopLevelStatements
+        //                : file.Classes.First().Methods.First(x => x.Name == "Main");
+
+        //            var statementToInsertBelow = targetBlock.Statements.LastOrDefault(x => x.HasMetadata("is-host-builder-configuration-statement")) ??
+        //                                         targetBlock.FindStatement(x => x.HasMetadata("is-builder-statement"));
+
+        //            statementToInsertBelow.InsertBelow(new CSharpInvocationStatement($"builder.Host.{methodName}")
+        //                .AddArgument(lambda)
+        //                .SeparatedFromPrevious()
+        //                .AddMetadata("is-host-builder-configuration-statement", true));
+        //        }, 10);
+        //    }
+
+        //    return this;
+        //}
+
+        public IProgramFile ConfigureHostBuilderChainStatement(
+            string methodName,
+            IEnumerable<string> parameters,
+            HostBuilderChainStatementConfiguration configure = null,
+            int priority = 0)
+        {
+            var parametersAsArray = parameters.ToArray();
+
+            if (!UsesMinimalHostingModel)
+            {
+                var hostBuilder = (IHasCSharpStatements)_cSharpFile.TopLevelStatements?.FindMethod("CreateHostBuilder") ??
+                                  _cSharpFile.Classes.First().FindMethod("CreateHostBuilder");
+
+                var hostBuilderChain = (CSharpMethodChainStatement)hostBuilder.Statements.First();
+                var appConfigurationBlock = (CSharpInvocationStatement)hostBuilderChain
+                    .FindStatement(stmt => stmt.ToString()!.StartsWith(methodName));
+                var lambda = (EditableCSharpLambdaBlock)appConfigurationBlock?.Statements.First();
+
+                if (appConfigurationBlock == null)
+                {
+                    lambda = new EditableCSharpLambdaBlock("()");
+
+                    appConfigurationBlock = new CSharpInvocationStatement(methodName)
+                        .WithoutSemicolon()
+                        .AddArgument(lambda);
+                    hostBuilderChain.Statements.Last()
+                        .InsertAbove(appConfigurationBlock);
+                }
+
+                SetParameters(lambda, parametersAsArray);
+                configure?.Invoke(lambda, (IReadOnlyList<string>)lambda.Metadata["parameters"]);
+            }
+            else
+            {
+                _cSharpFile.OnBuild(file =>
+                {
+                    var targetBlock = _usesTopLevelStatements
+                        ? (IHasCSharpStatements)file.TopLevelStatements
+                        : file.Classes.First().Methods.First(x => x.Name == "Main");
+
+                    var appConfigurationBlock = targetBlock.Statements
+                        .OfType<CSharpInvocationStatement>()
+                        .FirstOrDefault(x => x.ToString()!.TrimStart().StartsWith($"builder.Host.{methodName}("));
+                    var lambda = (EditableCSharpLambdaBlock)appConfigurationBlock?.Statements.First();
+
+                    if (appConfigurationBlock == null)
+                    {
+                        lambda = new EditableCSharpLambdaBlock("()");
+
+                        var statementToInsertBelow = targetBlock.Statements.LastOrDefault(x => x.HasMetadata("is-host-builder-configuration-statement")) ??
+                                                     targetBlock.FindStatement(x => x.HasMetadata("is-builder-statement"));
+
+                        statementToInsertBelow.InsertBelow(new CSharpInvocationStatement($"builder.Host.{methodName}")
+                            .AddArgument(lambda)
+                            .SeparatedFromPrevious()
+                            .AddMetadata("is-host-builder-configuration-statement", true));
+                    }
+
+                    SetParameters(lambda, parametersAsArray);
+                    configure?.Invoke(lambda, (IReadOnlyList<string>)lambda.Metadata["parameters"]);
+                });
+            }
+
+            return this;
+
+            static void SetParameters(EditableCSharpLambdaBlock lambda, IReadOnlyList<string> requestedParameters)
+            {
+                if (!lambda.TryGetMetadata<List<string>>("parameters", out var parameters))
+                {
+                    parameters = new List<string>();
+                    lambda.AddMetadata("parameters", parameters);
+                }
+
+                if (parameters.Count >= requestedParameters.Count)
+                {
+                    return;
+                }
+
+                parameters.AddRange(requestedParameters.Skip(parameters.Count));
+
+                lambda.UpdateText(parameters.Count == 1
+                    ? parameters[0]
+                    : $"({string.Join(", ", parameters)})");
+            }
+        }
+
+        public IProgramFile AddHostBuilderConfigurationStatement<TStatement>(
+            TStatement statement,
+            Action<TStatement> configure = null,
+            int priority = 0) where TStatement : CSharpStatement
+        {
+            if (!UsesMinimalHostingModel)
+            {
+                throw new InvalidOperationException("Only allowed when using minimal hosting model");
+            }
+
+            _cSharpFile.OnBuild(file =>
+            {
+
+                var targetBlock = _usesTopLevelStatements
+                    ? (IHasCSharpStatements)file.TopLevelStatements
+                    : file.Classes.First().Methods.First(x => x.Name == "Main");
+
+                var statementToInsertBelow = targetBlock.Statements.LastOrDefault(x => x.HasMetadata("is-host-builder-configuration-statement")) ??
+                                             targetBlock.FindStatement(x => x.HasMetadata("is-builder-statement"));
+
+                statementToInsertBelow.InsertBelow(statement);
+                statement.AddMetadata("is-host-builder-configuration-statement", true);
+                configure?.Invoke(statement);
+            });
+
+            return this;
+        }
+
+        public IProgramFile ConfigureMainStatementsBlock(Action<IHasCSharpStatements> configure)
+        {
+            _cSharpFile.OnBuild(file =>
+            {
+                var targetBlock = _usesTopLevelStatements
+                    ? (IHasCSharpStatements)file.TopLevelStatements
+                    : file.Classes.First().Methods.First(x => x.Name == "Main");
+
+                configure(targetBlock);
+            });
+
+            return this;
+        }
+
+        public IProgramFile AddMethod(
+            string returnType,
+            string name,
+            Action<IStartupMethod> configure = null,
+            int priority = 0)
+        {
+            StartupMethod.CreateOn(_template, returnType, name, configure, priority);
+            return this;
+        }
+
+        private class EditableCSharpLambdaBlock : CSharpLambdaBlock
+        {
+            public EditableCSharpLambdaBlock(string invocation) : base(invocation) { }
+
+            public void UpdateText(string text) => Text = text;
+        }
+    }
+}